+++ parser/element.cc
@@ -85,6 +85,7 @@ void Element::PublishMembers()
   publish("phase",     &Element::phase);
   publish("tOffset",   &Element::tOffset);
   publish("fieldModulator", &Element::fieldModulator);
+  publish("kg",        &Element::kg);
 
   // rmatrix elements, only 4x4
   publish("kick1",     &Element::kick1);
@@ -213,6 +214,14 @@ void Element::PublishMembers()
   publish("undulatorPeriod",       &Element::undulatorPeriod);
   publish("undulatorGap",          &Element::undulatorGap);
   publish("undulatorMagnetHeight", &Element::undulatorMagnetHeight);
+
+  // for gabor lens
+  publish("anodeRadius",         &Element::anodeRadius);
+  publish("anodeLength",         &Element::anodeLength);
+  publish("anodeThickness",      &Element::anodeThickness);
+  publish("electrodeRadius",     &Element::electrodeRadius);
+  publish("electrodeLength",     &Element::electrodeLength);
+  publish("electrodeThickness",  &Element::electrodeThickness);
 
   // bias
   publish("bias",                &Element::bias);
@@ -407,6 +416,12 @@ void Element::print(int ident) const
                   << "psi=   " << psi   << "rad" << std::endl;
         break;
       }
+    case ElementType::_GABORLENS:
+      {
+        std::cout << "b=  " << B  << "T" << std::endl
+                  << "kg=  " << kg  << std::endl;
+        break;
+      }
     default:
       {break;}
     }
@@ -483,6 +498,7 @@ void Element::flush()
   phase     = 0;
   tOffset   = 0;
   fieldModulator = "";
+  kg = 0;
 
   // rmatrix
   kick1 = 0;
@@ -590,6 +606,14 @@ void Element::flush()
   undulatorPeriod = 1;
   undulatorGap = 0;
   undulatorMagnetHeight = 0;
+
+  // gabor lens
+  anodeLength = 0;
+  anodeRadius = 0;
+  anodeThickness = 0;
+  electrodeLength = 0;
+  electrodeRadius = 0;
+  electrodeThickness = 0;
 
   // bias
   bias         = "";