+++ parser/elementtype.h
@@ -73,7 +73,8 @@ namespace GMAD
     _CT            = 75,
     _TARGET        = 76,
     _RFX           = 77,
-    _RFY           = 78
+    _RFY           = 78,
+    _GABORLENS     = 79
   };
 
   /// conversion from enum to string